Add tree data structure
The Tree struct may or may not be useful in the future. At the moment it
will be used to store information that the nodes should not, such as
initial board state (the nodes store moves instead of entire boards).

node_free will free the node after recursively freeing its children.
This allows you to cut of branches.

minimax takes the tree root as a node pointer but can also take NULL.

Also fixed a minimax mem leak: moves was not being freed if we returned
early.