Optimise searching for a king's checker.
The function turns the king into each piece type and then searches for
legal moves. We then look at the target square of each move to see if
there is an enemy piece on that square of relevant type. If there is, we
know that piece can put check on the king.

Time taken for a depth 6 search on an initialised board with the
new function: 32.28
old function: 224.18
Time taken for a depth 5 search with the
new function: 2.04
old function: 14.59

A test was written for each piece type to make sure they can all
correctly check the king.