Add precomputation during player's turn.
minimax_sub_boards_async takes the board state before the player's move,
constructs an array of trees (each of which represents the board after
the player makes a move) and calls minimax_sub_boards in a new thread
before returning the id of the created thread.

After the AI sends its move to the client, we call the above function.
When the player moves we check to see if we have already precomputed a
response or if we are currently searching the correct response.

g_player_moved and g_cancel_search are two globals that help control the
async thread and prevent needless computation.