Implement AI forced mate
Make the AI force a checkmate when it knows it can. Previously, it would
continuously pick a line to checkmate that never shrunk in depth. For
example, it would pick a mate that took 3 moves, every time it is its
turn. Now, it will pick a mate that takes 3 moves, on the next turn it
will pick a mate that takes 2 moves, then 1. This is implemented by
storing a checkmate line and executing that line.